fix: update to rust nightly

+++ benches/macro.rs
@@ -11,6 +11,6 @@ fn fib5(b: &mut Bencher) {
     b.iter(|| fib(5))
 }
 
-fn fib(n: uint) -> uint {
+fn fib(n: usize) -> usize {
     if n < 2 { n } else { fib(n - 1) + fib(n - 2) }
 }
